Add documentation on using IntegrationHook.
This adds a doc page on IntegrationHook. It touches upon the purpose of
integrations, linking to the topic page on integrations and the guide on
writing them. It then provides a full example of an integration and
usage of a hook that results in an HTTP POST when review requests are
published.

Testing Done:
Built and read the docs. Checked for errors, broken links, and spelling
issues.

Reviewed at https://reviews.reviewboard.org/r/9374/