Redo the index page for the API docs for better guidance/organization.
The index page for the API documentation was nothing but a giant
toctree, which didn't really provide any guidance at all. This has been
replaced with a new index page that gives a brief overview of the API,
followed by sections on:

* Getting started with the API (containing the overview, authentication,
  and glossary)
* API bindings available (linking to the RBTools Python API docs)
* Working with special data (currently consisting of rich text fields
  and, later, extra_data handling)
* Available resources (linking to the resource tree and categories)
* Errors (linking to the full list of errors)

It also gets rid of any mention of the old API 1.0 docs, which isn't
worth including anymore.

Testing Done:
Checked for build and spelling errors.

Tested that the links pointed to the right places.

Reviewed at https://reviews.reviewboard.org/r/9377/