Fix the documentation for APIExtraDataAccessHook.
The documentation for `APIExtraDataAccessHook` had some bad references,
syntax errors, and examples that didn't depict the actual usage of the
hook. The inline code docs for the hook also showed the wrong usage.

This fixes up all the docs to reflect the real usage of the hook and to
clean up the errors so that everything is presented and linked
correctly.

Testing Done:
Built the docs. Checked all links and checked that rendering issues
were fixed.

Reviewed at https://reviews.reviewboard.org/r/9382/