Feature-gate the new class-based actions system and shutdown hooks correctly
Since https://reviews.reviewboard.org/r/9131/ did not make it into the
release window for RB 3.0, we cannot address the issues in the original
implementation of class-based actions. So we are gating them behind a
feature so they can be marked as experimental (and not to be used in
production). That way we will be able to fix up the issues in the
implementation for a future release.

This patch also addresses the issue of the new action hooks not
unregistering their child actions.

Documentation on action hooks has been reverted to how it was in
Review Board 2.5.

Testing Done:
- Unit tests all pass.
- Wrote a test extension which (1) registered a hook that uses
  class-based actions and (2) registered a hook using dict-based
  actions. The former were not available (and was logged as such) and
  the latter were.

Reviewed at https://reviews.reviewboard.org/r/9384/