Add documentation on working with extra_data in the API.
This adds a new page that covers how to store extra_data on resources
using the old assignment method and the newer merging/patching methods.
It covers usage in detail and links to the specs for more information.
It also talks about access restrictions, including the prefix for
private keys and the impact of `APIExtraDataAccessHook`.

All resources that use extra_data now link to the new docs. Some of the
older resources that talked about assignment have now had that
information removed in favor of the new page.

Testing Done:
Built and read through the new docs. Checked for build issues, spelling
mistakes, and all links.

Reviewed at https://reviews.reviewboard.org/r/9381/