Fix the Review.objects._query method when querying the current user.
The query returned by `Review.objects.from_user()` was behaving
strangely when called with the currently logged-in user, because it was
OR-ing in an additional query on top of the ones that are already done
(so that it additionally included reviews that were replies). This is
not what people expect from this query, which is used only in the
"Reviews" list on the user page and in a couple extensions.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/9411/