Fix a regression with searching for closed review requests.
When the search API was rewritten, it was updated to use `public()` in
order to be more secure. This ended up using the default of only
returning opened review requests, excluding any closed ones, regressing
the ability to match those in the quick search results.

This changes that query to not limit results to open review requests,
and adds unit tests to ensure this doesn't regress.

Testing Done:
Unit tests pass. They failed prior to the fix.

Reviewed at https://reviews.reviewboard.org/r/9424/