Add capability flags for Git symlinks and JSON patching to the API.
Git symlinks and JSON patching are two new features that clients of the
API should be able to check for prior to use. Review Board 3.0 didn't
ship with any capabilities for this, which is something worth fixing for
3.0.1. This change adds those capabilities.

It also enhances the tests around this. Our previous capabilities tests
weren't very complete, and the rest of the server information was not
checked. The unit tests now check that the returned results from the API
exactly match what we expect, based on the server's settings and the
capabilities dictionary.

Testing Done:
Saw the capabilities in the API payload.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9420/