Fix problems when a user closes their own issues requiring verification.
This change makes it so that users who click "Fixed" (or "Drop") on
issues that they created will transition the issue state directly to
closed, even if the issue had been set to require verification. As it
currently exists, the issue would transition to waiting for
verification, which was unexpected. This also fixes up an issue where a
user object was compared to a pk.

Testing done:
- Ran unit tests.
- Exercised the test case listed in the bug.

Fixes bug 4610.

Reviewed at https://reviews.reviewboard.org/r/9447/