Fix a bug where "Download Diff" would vanish.
When loading a diff, the "Download Diff" action would be visible for a
short time, and then vanish. The culprit was that we were setting the
`canDownloadDiff` attribute in the code that loads a new revision, but
we had optimized the first load a little bit by including the context in
the initial page payload rather than making a separate request.

This change fixes things so that we compute the correct state of this
inside the `parse` method, which is used for both the initial load as
well as subsequent revisions.

Testing done:
- Verified that "Download Diff" was no longer vanishing when it
  shouldn't.
- Ran js-tests.

Fixes bug 4615.

Reviewed at https://reviews.reviewboard.org/r/9449/