Fix checking for review request updates on the review request page.
The model/view split for the review request page broke checking for
updates to the review request. It still worked on the diff viewer, but
not on the review request updates page. The reason is that this page
does not provide a type of updates to check for, leaving it at the
default of `null`, which was not supported. Before the split, the
default would have been `undefined`, which was supported.

This change simply fixes the check, allowing any falsy values to be
considered valid to enable checking for all updates.

Testing Done:
Made some updates to a review request. After a bit, the update bubble
showed up.

Reviewed at https://reviews.reviewboard.org/r/9452/