Fix the display of Emojis in diffed text fields.
The diffs shown for text fields on a review request were leaving Emojis
at their native size of 64x64, which is too big. That was due to the
width being overridden to `auto` for images in those fields.

This change fixes that by adding an explicit height for Emoji images in
the base rule, and allowing the width to remain `auto`.

An `&:extend` LessCSS rule was also removed, as it wasn't really needed.

Testing Done:
Checked that Emojis were looking correct in review request fields and in
the diffed content.

Reviewed at https://reviews.reviewboard.org/r/9455/