Prevent a translation from injecting HTML in change entry labels.
The "Fix it!" text for the change entry label was being injected
directly into a jQuery object, which could allow a translation to
accidentally cause unescaped HTML characters to mess up the page. This
is a simple fix that ensures we're creating a text node.

Testing Done:
Tested that the text still showed up.

Reviewed at https://reviews.reviewboard.org/r/9459/