Fix positions when parsing diff fragment and page update payloads.
The new changes for binary payloads for comment diff fragments and page
update payloads attempted to increment the current position by the
parsed next position, instead of setting to that position. This meant
that multiple payloads in a single response resulted in corrupted data
and crashes.

The fix is simply to set the position instead of increment by it.

Along with this, the payload for review request update payloads now
specifies an encoding of utf-8.

Testing Done:
New unit tests pass.

Tested manually with a review request that was failing to load
diff comment fragments.

Reviewed at https://reviews.reviewboard.org/r/9468/