Refactor user file attachments URL computation to handle local sites.
The user file attachments resource URL used for uploading images via
drag-and-drop wasn't correct when a local site was present, because we
were computing it inside the JavaScript. This change makes it so we
include the URL as part of the `UserSession` model, which makes it work
correctly.

Testing done:
- Verified that I could now DnD images into a review request within a
  local site.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/9470/