Don't attempt to run tests from the common SVN tests class.
In order to run the same tests across both PySVN and Subvertpy, we keep
all our tests in a common class and then subclass that for each backend.
However, the test runner still sees this as a valid test class, given
the test functions in it.

This change fixes that by indicating that the tests class should be
private by prefixing the name with an underscore.  That tells the test
runner to ignore it, without having to set `__test__ = False` in the
class and `__test__ = True` in the subclasses.

Testing Done:
Ran the unit tests. Verified tests for both backends had run, but that
no tests from the common class were run (previously they'd run and then
show up as skipped).

Reviewed at https://reviews.reviewboard.org/r/9475/