Update some dependencies to be Python-specific.
We have a couple dependencies that no longer work on older versions of
Python, which would make installation problematic for users and CI
maintenance more troublesome. We're also missing a dependency for Python
2.6 and lower.

This change adds the `argparse` dependency for Python 2.6 and lower, and
caps the `paramiko` dependency to the latest version for each Python
version supported. It also changes the `mercurial` dependency in
`dev-requirements.txt` to choose an appropriate version for each version
of Python, using version specifiers.

Testing Done:
Manually tested installation with Python 2.6 and 2.7.

Reviewed at https://reviews.reviewboard.org/r/9526/