Fix display of change description draft text.
When I was reworking the various fields and editors, I fixed up some of
the naming so that we were no longer calling the change description text
the "close description". Unfortunately, in the process, I messed up some
of the plumbing, and the rendered text wasn't displayed on the initial
page load (it would still show up when the editor was opened, and stay
visible after it was closed).

This change fixes up the plumbing so that the text is set into the
review request editor model, therefore showing up in the UI.

Testing done:
Set a change description and then reloaded the page. Saw the text show
up, where before it was hidden until the editor was opened.

Reviewed at https://reviews.reviewboard.org/r/9608/