Don't reserve space for review headers that don't exist.
If a review is published without any header text, we still show a
"Comment" button there (where "Reply" would be if the header existed).
In this case, we had a bunch of extra padding around the empty header
element, which was creating a pretty ugly blank space. To fix this, I've
added an extra class in the case that the header text doesn't exist, and
removed the padding for that class.

Testing done:
Published a review with comments but no header text. Saw that the
"Comment" button was now positioned in a reasonable location, rather
than with a ton of whitespace above it.

Reviewed at https://reviews.reviewboard.org/r/9610/