Add documentation on configuring CodeCommit and Bitbucket Server.
This adds new doc entries in the Review Board manual for configuring
CodeCommit and Bitbucket Server. These are linked to in the main table
of contents for the administration guide, in the Git documentation, and
in the list of integrations on the website.

This follows the same basic template used for other services bundled by
Power Pack.

Testing Done:
Built the docs. Checked the links and content for errors.

Reviewed at https://reviews.reviewboard.org/r/9639/