Move review request status/timestamp info out of the template.
We've previously been generating the human-readable information on the
review request's status (created time, updated time) in the review
request box template. This was a bit messy, and covered up issues like
the created timestamp never showing up when there's a draft.

This logic has been moved server-side to a reusable function that allows
additional information (like the diffset information in the diff viewer)
to be tacked on. Other views could potentially add their own information
as well. This handles creating a consistent representation for each of
the bits of status, reducing the complexity and improving the
performance of the template in the process.

Since this is more consolidated, it also allows for new unit tests that
ensure the correct output.

Testing Done:
Unit tests pass.

Tested various states manually in the review request page and the diff
viewer.

Reviewed at https://reviews.reviewboard.org/r/9628/