Improve the documentation for TemplateHook.
This fleshes out some of the TemplateHook docs, going into more detail
on a few hook points, along with linking to other extension pages that
are relevant for them. It also moves from a table-based display of names
to a definition-based display, which helps keep longer descriptions
readable.

The example has been updated to make use of alternate template hook
points, in order to avoid recommending its usage for custom CSS or
JavaScript (which static media bundles are a better solution for).

This is in preparation for some new updates to the list of available
template hook points that will be coming soon.

Testing Done:
Built the docs and looked through them, checking for formatting and
linking errors.

Reviewed at https://reviews.reviewboard.org/r/9629/