Fix several build errors and warnings during doc building.
We had a few syntax errors, bad links, redundant references, bad
database values, and incorrect configuration settings for the
documentation that led to error output when building the Review Board
manual, and a handful of harmless but annoying warnings due to missing
image state when simulating data for the API documentation. This made it
hard at times to determine whether new changes to the documentation were
correct, particularly for new contributors to the documentation.

This change fixes the errors within the documentation and the
configuration options for the manual, and disables the logger that's
complaining about missing image files (which don't otherwise impact us).
It also updates us to Sphinx 1.7, which solves a couple of other
warnings with references and unveiled others.

There is one error remaining, which has to do with how Sphinx is
attempting to introspect the webapi resource registry in a way that's
causing an attribute error. This must be addressed in Djblets.

Testing Done:
Documentation built without errors (except for the registry one) on
Sphinx 1.7.

Reviewed at https://reviews.reviewboard.org/r/9657/