Fix user selector widget to cope with missing avatars.
This change fixes up a small bug in the user selector widget used in the
admin UI to not crash when users are missing avatars.

Testing done:
Configured a user without an avatar and was able to add them to a group.

Fixes bug 4655.

Reviewed at https://reviews.reviewboard.org/r/9688/