Fix help text for Media URL and Static URL.
The help text in the admin UI for the Media URL and Static URL fields
suggested that you could leave them blank to use the defaults. This
was a lie. This change fixes that up, and makes the fields required,
since they are.

Testing done:
Loaded the general settings page and saw the correct help text.

Reviewed at https://reviews.reviewboard.org/r/9696/