Fix showing the review request page when there's only a draft diff.
The recent work to centralize how the status line for a review request
is computed introduced a regression when viewing the diff viewer for a
brand new review request. If there's a draft diff but no existing
published diffs, we'd attempt to look up an attribute on a ``None``
diffset, causing a page crash.

We now build information on the latest published diff only if there is
one.

Testing Done:
Successfully opened the diff viewer without a crash on a brand new
review request.

Reviewed at https://reviews.reviewboard.org/r/9685/