Don't crash when fetching the first commit in a mercurial repo.
The first changeset cannot have a parent so the parents list would be empty,
causing a crash. This change adds a check for that.

Sometimes mercurial diffs use "0000000000000000000000000000000000000000"
to indicate that the parent is the "null" changeset, which does function
correctly.

Testing Done:
Used a repo with a few changesets. Saw that it will result into an error if
first changeset will be fetched with cmdline or hgweb.

Added this change and saw that cmdline and hgweb won't get an error.

Reviewed at https://reviews.reviewboard.org/r/9679/